requirements.txt
flask-sqlalchemy
flask-migrate
python-dotenv
requests